Incorporated constants, used constants to simplify rendering
Add a folder for constants. These constants include technical specs for the bluetooth transceiver, the characters sent to the Arduino, and an array of data for the motors. Used the motor array to loop and render all of the motors in a more concise way than the previously hardcoded manner. This loop is found in the rendering of MotorScreen.