Update BLE connection method
Change how BLE device is connected. Created bleConnector which is a singleton class that creates a device and provides a method for sending data to it. The device itself is no longer exposed and you do not need to store it in a seperate global variable. It stays in the bleConnector file as a local variable and its methods are exposed through the object.